requirements.txt
asgiref==3.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.10
django-classy-tags==2.0.0
django-cms==3.7.4
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-templated-mail==1.1.1
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
easy-thumbnails==2.7
html5lib==1.1
oauthlib==3.1.0
Pillow==7.2.0
pycparser==2.20
pytz==2020.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
Unidecode==1.1.1
webencodings==0.5.1
